week3/8_semantic_search/requirements.txt
openai
python-dotenv
scipy
numpy